Do less work for the common case of a Binary field.
Writing a Binary field is far more common than the case where we are flattening
a tuple of primitive values into a comma-separated string. Previously, we were
converting a Binary to a String and then back to a Binary. Now we will leave the
Binary alone and only work with Strings for the uncommon tuple-combining mode.